Note: listen_notify.coffee needs to be refactored to read connection string data from a config file
Got a lot of plumbing done. have server side script
that sets up the socket.io listener and db listener
next step is to refactor the db listener so that on
a recieved notification instead of console.log it
sends the notification to the socket for emit to the
client.

I also think that we should refactor so that the
client on connection will send the channel it wishes
to listen for.